requirements.txt
alembic==1.8.0
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
aws==0.2.5
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
binary==1.0.0
bleach==5.0.0
blinker==1.4
boto==2.49.0
camelot-py==0.10.1
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
chess==1.9.0
click==8.0.1
colorama==0.4.4
confuse==1.7.0
cryptography==36.0.1
cycler==0.10.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
distro==1.6.0
Django==3.2.7
django-taggit==2.1.0
djangorestframework==3.12.4
dnspython==2.1.0
ecdsa==0.17.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fabric==2.7.0
fastjsonschema==2.15.3
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
getmac==0.8.3
ghostscript==0.7
greenlet==1.1.2
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.2
ImageHash==4.2.1
invoke==1.7.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.2
lxml==4.7.1
Mako==1.2.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.21.2
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.3.3
pandas-profiling==2.9.0
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathlib2==2.3.7.post1
pdf2image==1.16.0
pdfminer==20191125
pdfminer.six==20211012
pdftopng==0.2.3
pgn==0.1.0
phik==0.12.2
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
prettytable==3.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.1
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.12.0
Pygments==2.12.0
PyHTML==1.3.1
PyJWT==2.4.0
pymongo==4.0.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.18.1
pytesseract==0.3.8
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.26.0
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.25
sqlparse==0.4.2
stack-data==0.3.0
svgwrite==1.4.1
tabula-py==2.3.0
tabulate==0.8.9
tangled-up-in-unicode==0.2.0
terminado==0.15.0
tinycss2==1.1.1
tk==0.1.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
Tree==0.2.4
urllib3==1.26.6
virtualenv==20.8.0
visions==0.5.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
widgetsnbextension==3.6.0
WTForms==2.3.3
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
